requirements.txt
# Python dependencies for DRAFTS pipeline
#
# Optional faster FITS I/O (binary wheels on Windows for Python 3.10–3.12):
#   pip install fitsio
# The pipeline uses Astropy when fitsio is not installed (e.g. Python 3.14 on Windows).
# Optional: faster FITS reads. Use with Python 3.10–3.12 on Windows (prebuilt wheels).
#
# fitsio
astropy
opencv-python
matplotlib
numba
numpy
pandas
psutil
pyyaml
scipy
seaborn
scikit-image
scikit-learn
timm
torch
torchvision
tqdm
your
blimpy
jplephem